TL/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
idna==2.10
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
